requirements.txt
fastapi~=0.116.1
uvicorn[standard]
pydantic[email]
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy>=2
alembic~=1.16.4
psycopg2-binary
pydantic-settings~=2.10.1